+++ CHANGELOG.md
@@ -10,6 +10,10 @@ Enhancements:
 - Rewrite the write operations for arrays to have much better performance.
 - Heavily improve the performance of write operations when a relative address is provided (the main module of the remote process is now cached).
 
+Bugfixes:
+
+- Properly read strings without a \0 delimiter instead of throwing an ArgumentOutOfRangeException.
+
 ## V1.1.1 (05 September 2016)
 
 Bugfixes: